requirements.txt
Flask
flask-cors
psycopg2-binary
bcrypt
PyJWT
gunicorn